requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
mysqlclient==2.0.3
django-cors-headers==3.10.0
gunicorn==20.1.0
whitenoise==5.3.0
